refactor(tracking): consolidate module constants in constants.py
Move COUNT_METRICS, RATIO_METRICS, TRACKER_ARRAY_COLS, HOTA_THRESHOLDS,
and report display/colour literals into constants.py. Re-export
COUNT_METRICS and OVERALL_LABEL from utils via __getattr__ for compatibility.
